requirements.txt
amqp==5.1.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.33.0
Django==4.2.3
django-decouple==2.1
django-htmx==1.15.0
djangorestframework==3.14.0
kombu==5.3.1
phonenumbers==8.13.15
prompt-toolkit==3.0.39
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
